requirements.txt
asgiref==3.8.1
astroid==3.3.10
dill==0.4.0
Django==5.2.3
et_xmlfile==2.0.0
isort==6.0.1
mccabe==0.7.0
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
tzdata==2025.2
